submodules/colmap-wrapper/requirements.txt
numpy
open3d
opencv-contrib-python
pyquaternion==0.9.9
matplotlib
tqdm==4.64.0
wget==3.2
pycolmap
pyexiftool